Refactor into modules and enhance UI
